requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

# Django
django>=4.0,<4.1
django-cors-headers==3.2.1
gunicorn>=20.1.0,<20.2

# DRF
djangorestframework==3.13.1
django-filter==2.2.0

# DB
celery==5.1.2
redis==3.5.3

# HPLC
typing_extensions>=4.3.0,<4.4.0
requests>=2.13.0,<2.14.0

# Utils
python-dateutil==2.8.1
pytz==2019.3
simplejson==3.17.2
markdown==3.4.1

